pyproject.toml
[tool.poetry]
name = "motoconnector"
version = "0.1.0"
description = ""
authors = ["Sobolevskiy <soboldez77@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.7"
django-phonenumber-field = "^7.2.0"
django-phonenumbers = "^1.0.1"
Pillow = "^10.1.0"
djangorestframework = "^3.14.0"
Markdown = "^3.5.1"
django-filter = "^23.3"
drf-yasg = "^1.21.7"
celery = "^5.3.5"
redis = "^5.0.1"
djangorestframework-simplejwt = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
